+++ src/main/java/com/island/content/plants/Grass.java
@@ -7,7 +7,7 @@ import com.island.content.SpeciesKey;
  * Grass implementation.
  */
 public class Grass extends Biomass {
-    public Grass(double maxBiomass, int speed) {
+    public Grass(long maxBiomass, int speed) {
         super("Grass", SpeciesKey.GRASS, maxBiomass, speed);
     }
 }
